requirements.txt
Flask==2.0.2
PyYAML==6.0
schedule==1.1.0
selenium==4.1.0
twilio==7.5.1
